Move AI ASSISTED text to chart
